docs: Update project structure and documentation
- Renamed main DM.py to lipid_risk_assessor.py for clarity
- Moved images to docs/images directory
- Archived old main.py in docs/archive
- Updated README.md to reflect new structure and details
- Added setuptools to requirements.txt
