--- 一级预防逻辑.txt
@@ -1,37 +1,0 @@
-一级预防
-
-第一步逻辑，使其符合以下条件之一时直接列为高危人群：
-
-1.LDL-C ≥ 4.9 mmol/L 或 TC ≥ 7.2 mmol/L；
-
-2.糖尿病患者且年龄 ≥ 40岁；
-
-3.CKD 3-4期。
-
-第二步逻辑：
-
-如果患者不符合第一步逻辑中的高危条件，则根据危险因素数量评估风险：（设置复选框，并设计数框）
-
-
-危险因素包括吸烟、低HDL—C（HDL-C<1.0 mmol/L (40 mg/d)）、年龄≥45/55岁（男性/女性）
-先计算危险因素总数，再结合有无高血压和血脂水平进行10年ASCVD发病风险
-
-if 无高血压，除了“4.1≤TC<5.2或2.6≤LDL-C<3.4”且合并3个危险因素，or “5.2≤TC<7.2或3.4≤LDL-C<4.9”且合并2-3个危险因素这两种情况为中危（5-9%），其余都为低危（<5%）
-
-if有高血压，0个危险因素， 为低危（5%）
-if有高血压and1个危险因素， 且“3.1≤TC<4.1或1.8≤LDL-C<2.6”为低危（<5%），其余均中危（5-9%）
-if有高血压and2个危险因素， 且“3.1≤TC<4.1或1.8≤LDL-C<2.6”为中危（5-9%）,其余为高危（≥10%）
-if有高血压and3个危险因素均为高危（≥10%）
-
-
-第三步逻辑：
-在第二步逻辑中，10年ASCVD发病危险为中危且年龄<55岁者，需进行下一步评估余生危险：（设置复选框，并设计数框）
-
-具有以下任意2个及以上危险因素者，定义为ASCVD高危人群：
-
-(1)收缩压≥160mmHg或舒张压≥100mmHg;
-(2)非HDL-C≥5.2 mmol/L(200 mg/dl)；
-(3)HDL-C<1.0 mmol/L (40 mg/d)；
-(4) BMI 28 kg/m2；
-(5)吸烟
-
